requirements.txt
odfpy==1.4.0
